fix: Fixed a bug that caused inaccurate ball-inside-polygon detection
i only iterated the segment start point list and the segment end point list while the moving trail is not included. So if i try to detect the triangle i drew, only two edges is going under the detection test while the moving trail(the third edge of the triangle) missed the test. So if the point inside polygon is depend on some edge as this edge is the moving trail edge, the result is wrong. same for the other polygon.